[TASK] use ContentObjectRenderer
to get container child records for FE

cleanup pageview layer
unused method in database
