+++ Classes/Domain/Factory/PageView/Backend/ContainerFactory.php
@@ -16,7 +16,7 @@ use B13\Container\Domain\Factory\Database;
 use B13\Container\Tca\Registry;
 use TYPO3\CMS\Core\Context\Context;
 
-class ContainerFactory extends \B13\Container\Domain\Factory\PageView\ContainerFactory
+class ContainerFactory extends \B13\Container\Domain\Factory\ContainerFactory
 {
     /**
      * @var ContentStorage
@@ -31,5 +31,37 @@ class ContainerFactory extends \B13\Container\Domain\Factory\PageView\ContainerF
     ) {
         parent::__construct($database, $tcaRegistry, $context);
         $this->contentStorage = $contentStorage;
+    }
+
+    protected function children(array $containerRecord, int $language): array
+    {
+        return $this->contentStorage->getContainerChildren($containerRecord, $language);
+    }
+
+    protected function localizedRecordsByDefaultRecords(array $defaultRecords, int $language): array
+    {
+        $childRecords = parent::localizedRecordsByDefaultRecords($defaultRecords, $language);
+        return $this->contentStorage->workspaceOverlay($childRecords);
+    }
+
+    protected function containerByUid(int $uid): ?array
+    {
+        $record =  $this->database->fetchOneRecord($uid);
+        if ($record === null) {
+            return null;
+        }
+        return $this->contentStorage->containerRecordWorkspaceOverlay($record);
+    }
+
+    protected function defaultContainer(array $localizedContainer): ?array
+    {
+        if (isset($localizedContainer['_ORIG_uid'])) {
+            $localizedContainer = $this->database->fetchOneRecord((int)$localizedContainer['uid']);
+        }
+        $defaultRecord = $this->database->fetchOneDefaultRecord($localizedContainer);
+        if ($defaultRecord === null) {
+            return null;
+        }
+        return $this->contentStorage->containerRecordWorkspaceOverlay($defaultRecord);
     }
 }
