+++ Tests/Functional/Frontend/Fixtures/Workspace/container_with_ws_child_moved_from_outside.csv
@@ -2,4 +2,4 @@
 ,"uid","pid","CType","header","colPos","tx_container_parent","t3ver_wsid","t3ver_oid","t3ver_state"
 ,200,1,"b13-2cols-with-header-container",,,,,,
 ,201,1,"header","header-live",0,0,0,,
-,202,1,"header","header-ws",200,200,1,201,0
+,202,1,"header","header-ws",200,200,1,201,4
