+++ Tests/Unit/Domain/Factory/PageView/Backend/ContainerFactoryTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace B13\Container\Tests\Unit\Domain\Factory\PageView;
+namespace B13\Container\Tests\Unit\Domain\Factory\PageView\Backend;
 
 /*
  * This file is part of TYPO3 CMS-based extension "container" by b13.
@@ -13,7 +13,8 @@ namespace B13\Container\Tests\Unit\Domain\Factory\PageView;
  */
 
 use B13\Container\Domain\Factory\Database;
-use B13\Container\Domain\Factory\PageView\ContainerFactory;
+use B13\Container\Domain\Factory\PageView\Backend\ContainerFactory;
+use B13\Container\Domain\Factory\PageView\Backend\ContentStorage;
 use B13\Container\Tca\Registry;
 use TYPO3\CMS\Core\Context\Context;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
@@ -34,8 +35,9 @@ class ContainerFactoryTest extends UnitTestCase
         $database->expects(self::once())->method('fetchOneRecord')->with(1)->willReturn(null);
         $tcaRegistry = $this->getMockBuilder(Registry::class)->disableOriginalConstructor()->getMock();
         $context = $this->getMockBuilder(Context::class)->getMock();
+        $contentStorage = $this->getMockBuilder(ContentStorage::class)->disableOriginalConstructor()->getMock();
         $containerFactory = $this->getMockBuilder($this->buildAccessibleProxy(ContainerFactory::class))
-            ->setConstructorArgs(['database' => $database, 'tcaRegistry' => $tcaRegistry, 'context' => $context])
+            ->setConstructorArgs(['database' => $database, 'tcaRegistry' => $tcaRegistry, 'context' => $context, 'contentStorage' => $contentStorage])
             ->onlyMethods([])
             ->getMock();
         self::assertNull($containerFactory->_call('containerByUid', 1));
